Standardised the MediaLinkedCase logic
